refactor: APIクライアントをDIパターンに基づいて再構成
- ApiClientインターフェースを定義し、契約を明確化
- FirebaseApiClientとMockApiClientを別クラスとして実装
- 環境に応じたクライアントをファクトリーパターンで提供
- テストがより書きやすく、拡張性の高い構造に改善

この変更により：
- 型安全性が向上
- テストでのモックがより柔軟に
- 将来的な新しいAPIプロバイダーの追加が容易に
- 使用側（コンポーネント）の変更は不要
